Create common.ts to avoid circular dependency
