+++ src/helpers/schemas/common.ts
@@ -1,0 +1,19 @@
+/**
+ * These are common schemas that are used by other schemas.
+ * It is moved here to avoid circular dependency.
+ */
+import * as z from "zod";
+
+import { idRegexCheck, idRegexErrorMessage } from "./helper";
+
+export const StudyIdSchema = z.string().refine(idRegexCheck, {
+  message: idRegexErrorMessage("Study ID"),
+});
+
+export const StreamNameSchema = z.string().refine(idRegexCheck, {
+  message: idRegexErrorMessage("Stream name"),
+});
+
+export const QuestionIdSchema = z.string().refine(idRegexCheck, {
+  message: idRegexErrorMessage("Question ID"),
+});
